requirements.txt
numpy>=1.24.0,<2.0.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
scipy>=1.11.0
streamlit>=1.28.0
matplotlib>=3.7.0
pillow>=10.0.0
joblib>=1.3.0
pandas>=2.0.0
tqdm>=4.65.0

# Deep learning dependencies
torch>=2.0.0
torchvision>=0.15.0
datasets>=2.14.0
huggingface_hub>=0.19.0

# Optional but recommended for advanced models
# timm>=0.9.0
# accelerate>=0.20.0
# albumentations>=1.3.0